Debug correspondences
Found a potentail bug in the correspondence between frames and 3D points, etc. Changed some vectors to maps to fix this discrepancy. Also found a bug related to the imagename mapping which is now resolved. Tested on videos and found video mapping to work sucessfully with the outputs made from the fusion program.