Fixed a subtle bug
Had a bug where the height/width info was being grabbed from empty objects. So I switched things around so that a dictionary is built with "frameMetadata" objects that hold that information separately from the objects that hold image_idx and traversal_depth, etc. Those objects have an update mechanism in place "next_data" variable that updates those fields, while height and width did not. So this fixes that problem. Tested on a full dev instance on a video and it seems to work fine.